# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/minimumspecification": {
            "get": {
                "description": "Get a list of MinimumSpecification.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MinimumSpecification"
                ],
                "summary": "Get all MinimumSpecification.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MinimumSpecification"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creating a new MinimumSpecification.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MinimumSpecification"
                ],
                "summary": "Create New MinimumSpecification.",
                "parameters": [
                    {
                        "description": "the body to create a new MinimumSpecification",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.minimumSpecificationInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MinimumSpecification"
                        }
                    }
                }
            }
        },
        "/minimumspecification/{id}": {
            "get": {
                "description": "Get an MinimumSpecification by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MinimumSpecification"
                ],
                "summary": "Get MinimumSpecification.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MinimumSpecification id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MinimumSpecification"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete a MinimumSpecification by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MinimumSpecification"
                ],
                "summary": "Delete one MinimumSpecification.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MinimumSpecification id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update MinimumSpecification by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MinimumSpecification"
                ],
                "summary": "Update MinimumSpecification.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MinimumSpecification id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update MinimumSpecification",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.minimumSpecificationInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MinimumSpecification"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.minimumSpecificationInput": {
            "type": "object",
            "properties": {
                "directx": {
                    "type": "string"
                },
                "game_id": {
                    "type": "integer"
                },
                "graphics": {
                    "type": "string"
                },
                "memory": {
                    "type": "string"
                },
                "network": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "processor": {
                    "type": "string"
                },
                "soundcard": {
                    "type": "string"
                },
                "storage": {
                    "type": "string"
                }
            }
        },
...
        "models.MinimumSpecification": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "directx": {
                    "type": "string"
                },
                "gameID": {
                    "type": "integer"
                },
                "graphics": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "memory": {
                    "type": "string"
                },
                "network": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "processor": {
                    "type": "string"
                },
                "soundcard": {
                    "type": "string"
                },
                "storage": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}